// client/src/entities/Employee/model/consts/validateEmployeeError.ts
export enum ValidateEmployeeError {
	INCORRECT_EMPLOYEE_DATA = "incorrect_employee_data",
	INCORRECT_BIRTHDATE = "incorrect_birthdate",
	INCORRECT_REGISTRATION_ADDRESS = "incorrect_registration_address",
	INCORRECT_PHONE_NUMBER = "incorrect_phone_number",
	INCORRECT_EMAIL = "incorrect_email",
	INCORRECT_TRANSPORT_INFO = "incorrect_transport_info",
	INCORRECT_CONTRACT_DATA = "incorrect_contract_data",
	INCORRECT_DOCUMENTS_DATA = "incorrect_documents_data",
	INCORRECT_PAYROLLS_DATA = "incorrect_payrolls_data",
	NO_DATA = "no_data",
	SERVER_ERROR = "server_error",
}